add example with manual engine configuration and camunda-bpm-assert
