+++ examples/engine-camunda-bpm-assert/src/test/resources/logback-test.xml
@@ -1,0 +1,18 @@
+<configuration>
+
+  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+    <!-- encoders are assigned the type
+         ch.qos.logback.classic.encoder.PatternLayoutEncoder by default -->
+    <encoder>
+      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+    </encoder>
+  </appender>
+
+  <root level="info">
+    <appender-ref ref="STDOUT" />
+  </root>
+  
+  <logger name="org.camunda.bpm.extension.junit5" level="debug"/>
+  <logger name="org.camunda.bpm.engine.cmd" level="info"/>
+  <logger name="org.camunda.bpm.engine.impl.persistence.entity" level="info"/>
+</configuration>
